Loads of changes...

+++ include/Overlay.h
@@ -21,7 +21,7 @@ namespace de
 				virtual ~Overlay();
 
 				bool handleEvents( const SDL_Event &_event );
-				bool logic( const Uint32 &_deltaticks, State* &_nextState, options &_options );
+				bool logic( const float &_deltaticks, State* &_nextState, options &_options );
 				void render();
